requirements.txt
openai==0.26.1
python-dotenv==0.21.1
requests==2.22.0
fastapi[all]
